logd: build breakage aosp-brillo-master @ 2508494

+++ logd/LogKlog.cpp
@@ -286,7 +286,7 @@ void LogKlog::calculateCorrection(const log_time &monotonic,
     memset(&tm, 0, sizeof(tm));
     tm.tm_isdst = -1;
     localtime_r(&now, &tm);
-    if ((tm.tm_gmtoff < 0) && ((-tm.tm_gmtoff) > real.tv_sec)) {
+    if ((tm.tm_gmtoff < 0) && ((-tm.tm_gmtoff) > (long)real.tv_sec)) {
         real = log_time::EPOCH;
     } else {
         real.tv_sec += tm.tm_gmtoff;
